added AddAmenity methods to HotelAmenitiesCrossManager and RoomAmenitiesCrossManager, made unittests for both
