+++ app/src/androidTest/java/com/xepicgamerzx/hotelier/HotelAmenityManagerTest.java
@@ -13,6 +13,7 @@ import com.xepicgamerzx.hotelier.objects.HotelAmenity;
 import com.xepicgamerzx.hotelier.objects.Hotel;
 import com.xepicgamerzx.hotelier.objects.HotelRoom;
 import com.xepicgamerzx.hotelier.objects.HotelAmenitiesEnum;
+import com.xepicgamerzx.hotelier.storage.HotelAmenitiesCrossManager;
 import com.xepicgamerzx.hotelier.storage.HotelAmenityManager;
 import com.xepicgamerzx.hotelier.storage.HotelManager;
 import com.xepicgamerzx.hotelier.storage.HotelierDatabase;
@@ -27,6 +28,8 @@ import org.junit.runner.RunWith;
 import java.math.BigDecimal;
 import java.time.ZoneId;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 @RunWith(AndroidJUnit4.class)
 public class HotelAmenityManagerTest {
@@ -36,6 +39,7 @@ public class HotelAmenityManagerTest {
     private HotelManager hotelManager;
     private RoomManager roomManager;
     private HotelAmenityManager hotelAmenityManager;
+    private HotelAmenitiesCrossManager hotelAmenitiesCrossManager;
     private final ZoneId zoneId = ZoneId.systemDefault();
     private final BigDecimal price = BigDecimal.valueOf(200.91);
     private final long startDate = System.currentTimeMillis();
@@ -65,6 +69,7 @@ public class HotelAmenityManagerTest {
         hotelManager = HotelManager.getManager(db);
         roomManager = RoomManager.getManager(db);
         hotelAmenityManager = HotelAmenityManager.getManager(db);
+        hotelAmenitiesCrossManager = HotelAmenitiesCrossManager.getManager(db);
 
         ArrayList<HotelRoom> rooms = new ArrayList<>();
 
@@ -97,10 +102,13 @@ public class HotelAmenityManagerTest {
         HotelAmenity amenity1 = hotelAmenityManager.create("Laundry");
         HotelAmenity amenity2 = hotelAmenityManager.create(HotelAmenitiesEnum.GYM);
 
-        hotelManager.addAmenityToHotel(testHotel, amenity1);
-        hotelManager.addAmenityToHotel(testHotel, amenity2);
+        hotelAmenitiesCrossManager.addAmenityToHotel(testHotel, amenity1);
+        hotelAmenitiesCrossManager.addAmenityToHotel(testHotel, amenity2);
 
-        // TODO : find out how to get a hotel's amenities? to finish this test
+        List<HotelAmenity> actual = hotelAmenitiesCrossManager.getRelated(testHotel);
+        List<HotelAmenity> expected = Arrays.asList(amenity2, amenity1);
+
+        assertEquals(actual, expected);
     }
 
 
@@ -109,6 +117,7 @@ public class HotelAmenityManagerTest {
         roomManager.close();
         hotelAmenityManager.close();
         hotelManager.close();
+        hotelAmenitiesCrossManager.close();
         db.close();
     }
