+++ app/src/androidTest/java/com/xepicgamerzx/hotelier/RoomAmenityManagerTest.java
@@ -14,6 +14,7 @@ import com.xepicgamerzx.hotelier.objects.RoomAmenitiesEnum;
 import com.xepicgamerzx.hotelier.objects.RoomAmenity;
 import com.xepicgamerzx.hotelier.storage.HotelManager;
 import com.xepicgamerzx.hotelier.storage.HotelierDatabase;
+import com.xepicgamerzx.hotelier.storage.RoomAmenitiesCrossManager;
 import com.xepicgamerzx.hotelier.storage.RoomAmenityManager;
 import com.xepicgamerzx.hotelier.storage.RoomManager;
 
@@ -26,6 +27,8 @@ import org.junit.runner.RunWith;
 import java.math.BigDecimal;
 import java.time.ZoneId;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
 @RunWith(AndroidJUnit4.class)
@@ -36,6 +39,7 @@ public class RoomAmenityManagerTest {
     private HotelManager hotelManager;
     private RoomManager roomManager;
     private RoomAmenityManager roomAmenityManager;
+    private RoomAmenitiesCrossManager roomAmenitiesCrossManager;
     private final ZoneId zoneId = ZoneId.systemDefault();
     private final BigDecimal price = BigDecimal.valueOf(200.91);
     private final long startDate = System.currentTimeMillis();
@@ -64,6 +68,7 @@ public class RoomAmenityManagerTest {
         hotelManager = HotelManager.getManager(db);
         roomManager = RoomManager.getManager(db);
         roomAmenityManager = RoomAmenityManager.getManager(db);
+        roomAmenitiesCrossManager = RoomAmenitiesCrossManager.getManager(db);
 
         ArrayList<HotelRoom> rooms = new ArrayList<>();
 
@@ -95,15 +100,15 @@ public class RoomAmenityManagerTest {
         RoomAmenity amenity1 = roomAmenityManager.create("Patio");
         RoomAmenity amenity2 = roomAmenityManager.create(RoomAmenitiesEnum.WIFI);
 
-        List<HotelRoom> rooms = roomManager.getAll();
+        HotelRoom room = roomManager.getAll().get(1);
 
-        roomManager.addAmenityToRoom(rooms.get(0), amenity1);
-        roomManager.addAmenityToRoom(rooms.get(0), amenity2);
-        roomManager.addAmenityToRoom(rooms.get(1), amenity1);
+        roomAmenitiesCrossManager.addAmenityToRoom(room, amenity1);
+        roomAmenitiesCrossManager.addAmenityToRoom(room, amenity2);
 
-        assert(roomAmenityManager.getAmenitiesInRoom(rooms.get(0)).contains(amenity1));
-        assert(roomAmenityManager.getAmenitiesInRoom(rooms.get(0)).contains(amenity2));
-        assert(roomAmenityManager.getAmenitiesInRoom(rooms.get(1)).contains(amenity1));
+        List <RoomAmenity> actual = roomAmenitiesCrossManager.getRelated(room);
+        List <RoomAmenity> expected = Arrays.asList(amenity1, amenity2);
+
+        assertEquals(actual, expected);
     }
 
     @After
@@ -111,6 +116,7 @@ public class RoomAmenityManagerTest {
         roomManager.close();
         roomAmenityManager.close();
         hotelManager.close();
+        roomAmenitiesCrossManager.close();
         db.close();
     }
