+++ app/src/main/java/com/xepicgamerzx/hotelier/storage/HotelManager.java
@@ -119,11 +119,6 @@ public class HotelManager implements DiscreteManager<Hotel, Long, Long[]> {
         hotelDao.update(hotel);
     }
 
-    @Deprecated // Move to hotel amenities cross manager
-    public void addAmenityToHotel(Hotel hotel, HotelAmenity hotelAmenity) {
-        HotelAmenitiesCrossRef hotelAmenitiesCrossRef = new HotelAmenitiesCrossRef(hotel, hotelAmenity);
-        hotelAmenitiesCrossDao.insert(hotelAmenitiesCrossRef);
-    }
 
     /**
      * Get hotels with matching hotel IDs.
