+++ app/src/main/java/com/xepicgamerzx/hotelier/storage/RoomManager.java
@@ -187,12 +187,6 @@ public class RoomManager implements DiscreteManager<HotelRoom, Long, Long[]> {
         return getPriceRange(hotelRooms);
     }
 
-    @Deprecated // move to room amenities cross manager
-    public void addAmenityToRoom(HotelRoom hotelRoom, RoomAmenity roomAmenity) {
-        RoomAmenitiesCrossRef roomAmenitiesCrossRef = new RoomAmenitiesCrossRef(hotelRoom, roomAmenity);
-        roomAmenitiesCrossDao.insert(roomAmenitiesCrossRef);
-    }
-
     /**
      * Closes the database if already open.
      */
